adds splitBy extension method to scala collections
`Iterator#splitBy` constructs an iterator where consecutive elements of
the original iterator are accumulated as long as the output of a key
function for each element doesn't change.

This operation makes sense as soon as you are trying to process an
iterator where you know the elements will be sorted in a certain way and
you need to group them without loading all the data in memory.

For instance
 * processing a file where the ordering is guaranteed but the file
 doesn't fit in the heap,
 * processing a streaming resultset where the underlying database
 guarantees the ordering because of a sort clause.

The same operation is added to `Iterable` with the difference that the
specific container type of the input is preserved for both collection
levels of the output, thus

* `Set(1,2,3).splitBy(identity)` returns
    `Set(Set(1), Set(2), Set(3))`
* `Vector(1,2,3).splitBy(identity)` returns
    `Vector(Vector1), Vector2), Vector3))`
* etc.